FerremasV3-main/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-cors-headers==4.3.1
django-rest-swagger==2.2.0
django_templates==0.0.13
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-yasg==1.21.10
fastapi==0.115.12
h11==0.16.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mysqlclient==2.2.7
openapi-codec==1.3.2
packaging==25.0
pillow==11.2.1
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.9.0
PyMySQL==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
ruamel.yaml==0.18.10
setuptools==80.3.1
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
